[dtensor] changes required before moving to core
[ghstack-poisoned]
