+++ test/spmd/tensor/parallel/test_tp_examples.py
@@ -3,8 +3,8 @@ import torch
 import torch.nn as nn
 import functools
 from torch.testing._internal.common_utils import run_tests
-from spmd.testing.common_utils import DistTensorTestBase, with_comms, NUM_DEVICES, skip_unless_torch_gpu  # type: ignore
-from spmd import (
+from spmd.testing.common_dtensor import DTensorTestBase, with_comms, NUM_DEVICES, skip_unless_torch_gpu  # type: ignore
+from spmd.tensor import (
     distribute_tensor,
     distribute_module,
     DeviceMesh,
@@ -122,7 +122,7 @@ class MultiheadAttnWrap(nn.Module):
         return self.attn(query, key, value)
 
 
-class DistTensorParallelExampleTest(DistTensorTestBase):
+class DistTensorParallelExampleTest(DTensorTestBase):
     @with_comms
     def test_mlp_megatron_e2e(self):
         inp_size = [5, 10]
