+++ test/spmd/tensor/parallel/test_view_sharding_dim_change.py
@@ -1,17 +1,17 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 import torch
 from torch.testing._internal.common_utils import run_tests
-from spmd.testing.common_utils import (  # type: ignore
-    DistTensorTestBase,
+from spmd.testing.common_dtensor import (  # type: ignore
+    DTensorTestBase,
     with_comms,
 )
-from spmd import DeviceMesh, DTensor, Shard
+from spmd.tensor import DeviceMesh, DTensor, Shard
 from spmd.tensor.parallel._view_with_dim_change import (
     _view_with_sharding_dim_change,
 )
 
 
-class TPViewShardingDimChangeTest(DistTensorTestBase):
+class TPViewShardingDimChangeTest(DTensorTestBase):
     @with_comms
     def test_view_with_sharding_dim_change(self):
         device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
