+++ test/spmd/tensor/test_common_rules.py
@@ -10,14 +10,14 @@ from spmd.tensor.ops.common_rules import (
     pointwise_rule,
 )
 from spmd.tensor.placement_types import DTensorSpec
-from spmd.testing.common_utils import (
-    DistTensorTestBase,
+from spmd.testing.common_dtensor import (
+    DTensorTestBase,
     with_comms,
 )
-from spmd import DeviceMesh
+from spmd.tensor import DeviceMesh
 
 
-class CommonRulesTest(DistTensorTestBase):
+class CommonRulesTest(DTensorTestBase):
     def parse_schema(self, schema_str):
         return FunctionSchema.parse(schema_str)
